+++ lib/amazonka-core/src/Amazonka/Types.hs
@@ -132,6 +132,7 @@ module Amazonka.Types
         Jakarta,
         Melbourne,
         Mumbai,
+        Hyderabad,
         Osaka,
         Seoul,
         Singapore,
@@ -880,6 +881,9 @@ pattern Melbourne = Region' "ap-southeast-4"
 pattern Mumbai :: Region
 pattern Mumbai = Region' "ap-south-1"
 
+pattern Hyderabad :: Region
+pattern Hyderabad = Region' "ap-south-2"
+
 pattern Osaka :: Region
 pattern Osaka = Region' "ap-northeast-3"
 
@@ -966,6 +970,7 @@ pattern Ningxia = Region' "cn-northwest-1"
   Jakarta,
   Melbourne,
   Mumbai,
+  Hyderabad,
   Osaka,
   Seoul,
   Singapore,